references/pyproject.toml
[build-system]
# https://setuptools.pypa.io/en/latest/build_meta.html
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
# https://setuptools.pypa.io/en/latest/userguide/pyproject_config.html
name = "acemagic_s1"
version = "0.1.0"
description = "simple fast-api app learning and getting real-time monitoring."
readme = "README.md"

authors = [{ name = "Daniel Ramirez", email = "dxas90@gmail.com" }]

requires-python = ">=3.10.0"

# required dependencies
dependencies = [
  "annotated-types==0.7.0",
  "anyio==4.7.0",
  "certifi==2024.8.30",
  "charset-normalizer==3.4.0",
  "docutils==0.21.2",
  "fastapi==0.115.6",
  "h11==0.14.0",
  "h2==4.1.0",
  "hpack==4.0.0",
  "Hypercorn==0.17.3",
  "hyperframe==6.0.1",
  "idna==3.10",
  "importlib_metadata==8.5.0",
  "libusb==1.0.27.post3",
  "packaging==24.2",
  "pkg_about==1.2.5",
  "priority==2.0.0",
  "psutil==6.1.0",
  "pydantic_core==2.27.1",
  "pydantic==2.10.3",
  "pyserial==3.5",
  "pyusb==1.2.1",
  "requests==2.32.3",
  "setuptools==75.6.0",
  "sniffio==1.3.1",
  "starlette==0.41.3",
  "toml==0.10.2",
  "typing_extensions==4.12.2",
  "urllib3==2.2.3",
  "wsproto==1.2.0",
  "zipp==3.21.0",
]

[project.optional-dependencies]
# recursive include for dev purposes
dev = ["black", "httpx", "pytest", "pytest-cov"]

[tool.setuptools.packages.find]
where = ["."]
include = ["micropulse"]
exclude = ["tests"]
